Refactor based on RubyCritic guidance
This is by no means comprehensive. I'm sure that it leaves a lot on the
table. I just looked through some of the RubyCritic feedback and made
some refactors that looked good/helpful to me.
