+++ app/controllers/application_controller.rb
@@ -2,7 +2,9 @@ class ApplicationController < ActionController::Base
   include Bootstrappable
   include BrowserSupportCheckable
   include Classable
+  include JsonPrioritizable
   include Pundit::Authorization
+  include Redirectable
   include RequestRecordable
   include SchemaValidatable
   include TokenAuthenticatable
@@ -15,11 +17,9 @@ class ApplicationController < ActionController::Base
   protect_from_forgery with: :exception
 
   before_action :authenticate_user!
-  before_action :prioritize_json_format
   before_action :set_paper_trail_whodunnit
   before_action :set_browser_uuid
   before_action :set_controller_action_in_context
-  before_action :store_redirect_chain
 
   after_action :verify_authorized, unless: :skip_authorization?
 
@@ -151,74 +151,5 @@ class ApplicationController < ActionController::Base
         render_json_error(message, 403)
       end
     end
-  end
-
-  # rubocop:disable Metrics/PerceivedComplexity
-  def prioritize_json_format
-    if request.accepts.any?
-      # Parse accept header directly since Mime::Type doesn't expose parameters
-      accepts = request.headers['Accept'].to_s.split(',').map do |accept|
-        type, params = accept.strip.split(';')
-        # rubocop:disable Lint/NumberConversion
-        q = params&.match(/q=([0-9.]+)/)&.[](1)&.to_f || 1.0
-        # rubocop:enable Lint/NumberConversion
-        [type, q]
-      end.sort_by { |_, q| -q } # Sort by q value descending
-
-      # If application/json has highest priority (tied or better),
-      # force request format to JSON
-      json_priority = accepts.find { |type, _| type == 'application/json' }&.last || 0
-      highest_priority = accepts.first&.last || 0
-
-      if json_priority >= highest_priority && json_priority > 0
-        request.format = :json
-      end
-    end
-  end
-  # rubocop:enable Metrics/PerceivedComplexity
-
-  def redirect_location
-    next_redirect_chain_value ||
-      session.delete('user_return_to') ||
-      root_path
-  end
-
-  def store_redirect_chain
-    if (redirect_chain =
-          params[:redirect_chain] ||
-          request.env['omniauth.origin'].presence
-       ) && redirect_chain != new_user_session_url
-      session[:redirect_chain] = redirect_chain
-    end
-  end
-
-  def next_redirect_chain_value
-    if (redirect_chain = session.delete(:redirect_chain)&.split('|'))
-      next_value = first_redirect_chain_value_to_follow(redirect_chain)
-
-      if redirect_chain.present?
-        session[:redirect_chain] = redirect_chain.join('|')
-      end
-
-      next_value
-    end
-  end
-
-  def first_redirect_chain_value_to_follow(redirect_chain)
-    20.times do
-      next_value = redirect_chain.shift
-
-      if next_value.start_with?('wizard:')
-        wizard_type = next_value.delete_prefix('wizard:')
-
-        if wizard_type == 'set-public-name-if-new' && current_user.created_at > 1.minute.ago
-          return edit_public_name_my_account_path
-        end
-      else
-        return next_value
-      end
-    end
-
-    nil
   end
 end
