+++ app/workers/application_worker.rb
@@ -27,10 +27,12 @@ module ApplicationWorker
 
   def perform(*args)
     with_lock_or_reschedule(args) do
-      disable_flag_name = :"disable_#{self.class.name.underscore}_worker"
+      class_name = self.class.name
+      disable_flag_name = :"disable_#{class_name.underscore}_worker"
+
       if Flipper.enabled?(disable_flag_name)
         logger.info(<<~LOG.squish)
-          Skipping #{self.class.name} job because the `#{disable_flag_name}` flag is enabled.
+          Skipping #{class_name} job because the `#{disable_flag_name}` flag is enabled.
         LOG
         return
       end
